Fortran properties now computed with double precision

+++ ExternalMediaLibrary/branches/heliumlib/Projects/Heliumlib/heliumlib.h
@@ -2,10 +2,10 @@
    Header file with external declarations for FORTRAN function calls
 */
 
-extern "C" _declspec(dllimport) float DHE (float *P, float *T);
-extern "C" _declspec(dllimport) float HHE (float *P, float *T);
-extern "C" _declspec(dllimport) float CPHE (float *P, float *T);
-extern "C" _declspec(dllimport) float CVHE (float *P, float *T);
-extern "C" _declspec(dllimport) float GRUNHE (float *P, float *T);
-extern "C" _declspec(dllimport) float SNDHE (float *P, float *T);
-extern "C" _declspec(dllimport) float SHE (float *P, float *T);
+extern "C" _declspec(dllimport) double DHE (double *P, double *T);
+extern "C" _declspec(dllimport) double HHE (double *P, double *T);
+extern "C" _declspec(dllimport) double CPHE (double *P, double *T);
+extern "C" _declspec(dllimport) double CVHE (double *P, double *T);
+extern "C" _declspec(dllimport) double GRUNHE (double *P, double *T);
+extern "C" _declspec(dllimport) double SNDHE (double *P, double *T);
+extern "C" _declspec(dllimport) double SHE (double *P, double *T);
